Correction terminer

+++ app/Gestion/GestionPersonnel.php
@@ -40,6 +40,9 @@ class GestionPersonnel
 
 	public function delete($slug)
 	{
+		if (Auth::user()->login = $slug) {
+			return;
+		}
 		Personnel::whereLogin($slug)->delete();
 	}
 }
